refactor(data-platform): split repository ownership by domain
